IOS-00051 Foundation Models IOS-00051E Refactor - Actor and real data Models
