+++ TASKS.md
@@ -476,3 +476,7 @@
     B: ReadinessAnalysisFeature
     C: Framework Translation
     D: Refactor - AsyncStream
+    E: Refactor - Actor and real data
+    - DataAnalyzer
+    - HealthDataTool
+
