+++ WorkoutMirrorLive/FeaturesNew/StatsTab/Utilities/Client/DataAnalyzerClient.swift
@@ -35,33 +35,36 @@ extension DataAnalyzerClient: DependencyKey {
     
     // MARK: - Live Value
     
-    static let liveValue = DataAnalyzerClient(
-        isAvailable: {
-            #if canImport(FoundationModels)
-            if #available(iOS 26, *) {
-                return await DataAnalyzer.shared.available
-            }
-            #endif
-            return false
-        },
-        streamAnalysis: {
-            #if canImport(FoundationModels)
-            if #available(iOS 26, *) {
-                let available = await DataAnalyzer.shared.available
-                
-                if available {
-                    // AI jest - zwróć prawdziwy stream z modelu
-                    return await DataAnalyzer.shared.streamAnalysis()
-                } else {
-                    // AI nie ma - zwróć mock stream
-                    return mockStream()
+    static let liveValue: DataAnalyzerClient = {
+        #if canImport(FoundationModels)
+        if #available(iOS 26, *) {
+            let analyzer = DataAnalyzer()
+            
+            return DataAnalyzerClient(
+                isAvailable: {
+                    await analyzer.available
+                },
+                streamAnalysis: {
+                    let available = await analyzer.available
+                    
+                    if available {
+                        // AI available - return real stream
+                        return await analyzer.streamAnalysis()
+                    } else {
+                        // AI unavailable - return mock stream
+                        return mockStream()
+                    }
                 }
-            }
-            #endif
-            // iOS < 26 - zwróć mock stream
-            return mockStream()
+            )
         }
-    )
+        #endif
+        
+        // iOS < 26 - return mock
+        return DataAnalyzerClient(
+            isAvailable: { false },
+            streamAnalysis: { mockStream() }
+        )
+    }()
     
     // MARK: - Mock Stream Helper
