+++ WorkoutMirrorLive/FeaturesNew/StatsTab/Utilities/DataAnalyzer.swift
@@ -7,14 +7,11 @@
 
 import Foundation
 import FoundationModels
-import Observation
 
 @available(iOS 26, *)
-final class DataAnalyzer {
+actor DataAnalyzer {
     
-    static let shared = DataAnalyzer()
-    
-    let model: SystemLanguageModel = .default
+    private let model: SystemLanguageModel = .default
     
     var available: Bool {
         get async {
@@ -27,10 +24,10 @@ final class DataAnalyzer {
         }
     }
     
-    private init() {}
+    init() {}
     
     /// Returns AsyncStream of partial responses
-    func streamAnalysis() async -> AsyncStream<String> {
+    func streamAnalysis() -> AsyncStream<String> {
         return AsyncStream { continuation in
             Task {
                 await self.performAnalysis(continuation: continuation)
@@ -182,38 +179,6 @@ final class DataAnalyzer {
     }
 }
 
-// MARK: - Health Data Tool
-
-struct HealthDataTool: Tool {
-    var name: String = "fetchHealthMetrics"
-    var description: String = "Fetch the latest health metrics (RHR, HRV, sleep, activity, readiness)."
-    
-    @Generable()
-    struct Arguments {}
-    
-    func call(arguments: Arguments) async throws -> String {
-        let metrics = [
-            "restingHeartRate": 56,
-            "hrv": 95,
-            "sleepHours": 7.5,
-            "activityKcal": 750,
-            "trainingReadiness": 75
-//            "restingHeartRate": 72,      // ⬆️ Wysoki (normalnie ~58-60)
-//            "hrv": 45,                    // ⬇️ Niski (normalnie ~80-90)
-//            "sleepHours": 5.2,            // ⬇️ Za mało (normalnie 7-8h)
-//            "activityKcal": 1200,         // ⬆️ Ciężki trening wczoraj (normalnie ~650-750)
-//            "trainingReadiness": 28       // ❌ Bardzo słaby wynik
-        ] as [String : Any]
-        
-        return """
-        Resting Heart Rate: \(metrics["restingHeartRate"]!) bpm
-        HRV: \(metrics["hrv"]!) ms
-        Sleep: \(metrics["sleepHours"]!) hours
-        Activity: \(metrics["activityKcal"]!) kcal
-        Training Readiness Score: \(metrics["trainingReadiness"]!)
-        """
-    }
-}
 
 //@available(iOS 26, *)
 //#Playground {
